refactor(agent/agentcontext): collocate types with producers, rename watch to watcher
Move types.go contents next to their producers and consumers:

- Source moves into manager.go (Manager input).
- Resource, ResourceKind, ResourceStatus, their String() methods,
  Snapshot, ComputeAggregateHash, and writeLengthPrefixed move
  into resolve.go (the resolver emits and hashes these).
- types_test.go is folded into resolve_test.go.

Rename watch.go and watch_test.go to watcher.go and watcher_test.go;
the file describes the Watcher type.

Inline switchHomeEnv into paths_test.go (its only caller) and
delete helpers_test.go.

export_test.go stays. It is the Go-conventional bridge for
exposing the unexported started() channel to external test
packages and is recognized by the toolchain for this purpose.
