+++ agent/agentcontext/watcher_test.go
